refactor: component folders (#26)
