+++ next/components/BioNetworks/Network/Atlas/components/Overview/components/MainColumn/mainColumn.tsx
