+++ next/components/BioNetworks/Network/Atlas/components/Publications/components/MainColumn/mainColumn.tsx
