+++ next/components/BioNetworks/Network/Atlas/components/common/Tabs/tabs.tsx
