+++ next/components/BioNetworks/components/Index/components/MainColumn/mainColumn.styles.ts
