--- next/components/Network/components/common/Hero/hero.tsx
@@ -1,25 +1,0 @@
-import { BackPageHero } from "@clevercanary/data-explorer-ui/lib/components/Layout/components/BackPage/components/BackPageHero/backPageHero";
-import { NETWORKS_ROUTE } from "constants/routes";
-import { useNetwork } from "contexts/networkContext";
-import React, { useMemo } from "react";
-
-export const Hero = (): JSX.Element => {
-  const { network } = useNetwork();
-  const { name, path } = network;
-
-  const breadcrumbs = useMemo(
-    () => [
-      {
-        path: NETWORKS_ROUTE,
-        text: "Biological Networks",
-      },
-      {
-        path: NETWORKS_ROUTE + "/" + path,
-        text: name,
-      },
-    ],
-    [name, path]
-  );
-
-  return <BackPageHero title={name} breadcrumbs={breadcrumbs} />;
-};
