+++ next/pages/bio-networks/[network]/[atlas]/datasets.tsx
@@ -3,9 +3,9 @@ import { AtlasProvider } from "contexts/atlasContext";
 import { GetStaticPaths, GetStaticProps, InferGetStaticPropsType } from "next";
 import React from "react";
 import { AtlasContext } from "../../../../@types/network";
-import { Hero } from "../../../../components/Atlas/components/common/Hero/hero";
-import { Tabs } from "../../../../components/Atlas/components/common/Tabs/tabs";
-import { MainColumn } from "../../../../components/Atlas/components/Datasets/components/MainColumn/mainColumn";
+import { Hero } from "../../../../components/BioNetworks/Network/Atlas/components/common/Hero/hero";
+import { Tabs } from "../../../../components/BioNetworks/Network/Atlas/components/common/Tabs/tabs";
+import { MainColumn } from "../../../../components/BioNetworks/Network/Atlas/components/Datasets/components/MainColumn/mainColumn";
 import * as atlasPages from "../../../../utils/atlasPages";
 
 export const getStaticPaths: GetStaticPaths = atlasPages.getStaticPaths;
