+++ next/pages/bio-networks/[network]/[atlas]/publications.tsx
@@ -3,10 +3,10 @@ import { AtlasProvider } from "contexts/atlasContext";
 import { GetStaticPaths, GetStaticProps, InferGetStaticPropsType } from "next";
 import React from "react";
 import { AtlasContext } from "../../../../@types/network";
-import { Hero } from "../../../../components/Atlas/components/common/Hero/hero";
-import { Tabs } from "../../../../components/Atlas/components/common/Tabs/tabs";
-import { MainColumn } from "../../../../components/Atlas/components/Publications/components/MainColumn/mainColumn";
-import { SideColumn } from "../../../../components/Atlas/components/Publications/components/SideColumn/sideColumn";
+import { Hero } from "../../../../components/BioNetworks/Network/Atlas/components/common/Hero/hero";
+import { Tabs } from "../../../../components/BioNetworks/Network/Atlas/components/common/Tabs/tabs";
+import { MainColumn } from "../../../../components/BioNetworks/Network/Atlas/components/Publications/components/MainColumn/mainColumn";
+import { SideColumn } from "../../../../components/BioNetworks/Network/Atlas/components/Publications/components/SideColumn/sideColumn";
 import * as atlasPages from "../../../../utils/atlasPages";
 
 export const getStaticPaths: GetStaticPaths = atlasPages.getStaticPaths;
