+++ next/pages/bio-networks/[network]/index.tsx
@@ -1,10 +1,10 @@
 import { Detail } from "@clevercanary/data-explorer-ui/lib/components/Detail/detail";
 import { GetStaticPaths, GetStaticProps, InferGetStaticPropsType } from "next";
 import { NetworkParam } from "../../../@types/network";
-import { Hero } from "../../../components/Network/components/common/Hero/hero";
-import { Tabs } from "../../../components/Network/components/common/Tabs/tabs";
-import { MainColumn } from "../../../components/Network/components/Overview/components/MainColumn/mainColumn";
-import { SideColumn } from "../../../components/Network/components/Overview/components/SideColumn/sideColumn";
+import { Hero } from "../../../components/BioNetworks/Network/components/common/Hero/hero";
+import { Tabs } from "../../../components/BioNetworks/Network/components/common/Tabs/tabs";
+import { MainColumn } from "../../../components/BioNetworks/Network/components/Overview/components/MainColumn/mainColumn";
+import { SideColumn } from "../../../components/BioNetworks/Network/components/Overview/components/SideColumn/sideColumn";
 import { NetworkProvider } from "../../../contexts/networkContext";
 import * as networkPages from "../../../utils/networkPages";
